pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "Cython>=0.29"]
build-backend = "setuptools.build_meta"

[project]
name = "Owl_z" 
version = "0.1.0"
description = "A project for high-redshift galaxy and QSO modeling"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}  
authors = [
    {name = "Meriam Ezziati", email = "meezziati@gmail.com"}
]
keywords = ["astronomy", "cosmology", "bayesian", "quasars"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Astronomy",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "numpy>=1.20.0",
    "scipy>=1.7.0",
    "matplotlib>=3.3.0",
    "astropy>=4.3.0",
    "PyYAML>=5.4.0",
    "Cython>=0.29.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.12",
    "black>=21.0",
    "flake8>=3.9",
    "mypy>=0.910",
]

[project.urls]
Homepage = "https://github.com/MeriamEzziati-dot/Owl_z"
Repository = "https://github.com/MeriamEzziati-dot/Owl_z"
Issues = "https://github.com/MeriamEzziati-dot/Owl_z/issues"

[tool.setuptools]
packages = ["model", "colors"]

[tool.setuptools.package-data]
"*" = ["*.pyx", "*.pxd"]

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=model --cov=colors"
